api: break component compacting impl out into separate class
